Move pages to folder.
